sumrise/requirements.txt
feedparser==6.0.10
flask==2.3.1
openai-whisper==20230314
pip==23.2.1
pydub==0.25.1
requests==2.27.1
rq==1.15.1
schedule==1.2.0